Mve mockery to its own package
