+++ .gitignore
@@ -7,6 +7,7 @@
 
 # IDE
 .idea/
+oryxBuildBinary
 
 # Test binary, built with `go test -c`
 *.test
